Refactor: VerifyPasswordRequestModel as base class
This provides a common base class for all password/OTP verification
