+++ src/Core/Models/Api/Request/Accounts/VerifyPasswordRequestModel.cs
@@ -1,12 +1,25 @@
-﻿using System;
+﻿using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
 
 namespace Bit.Core.Models.Api
 {
-    public class VerifyPasswordRequestModel
+    public class VerifyPasswordRequestModel : IValidatableObject
     {
-        [Required]
         [StringLength(300)]
         public string MasterPasswordHash { get; set; }
+        public string OTP { get; set; }
+
+        public IEnumerable<ValidationResult> Validate(ValidationContext validationContext)
+        {
+            if (string.IsNullOrEmpty(MasterPasswordHash) && string.IsNullOrEmpty(OTP))
+            {
+                yield return new ValidationResult("MasterPasswordHash or OTP must be supplied.");
+            }
+        }
+
+        public bool SuppliedMasterPassword()
+        {
+            return !string.IsNullOrEmpty(MasterPasswordHash);
+        }
     }
 }
