+++ src/Core/Models/Api/Request/ApiKeyRequestModel.cs
@@ -1,24 +1,4 @@
-﻿using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations;
-
-namespace Bit.Core.Models.Api
+﻿namespace Bit.Core.Models.Api
 {
-    public class ApiKeyRequestModel : IValidatableObject
-    {
-        public string MasterPasswordHash { get; set; }
-        public string OTP { get; set; }
-
-        public IEnumerable<ValidationResult> Validate(ValidationContext validationContext)
-        {
-            if (string.IsNullOrEmpty(MasterPasswordHash) && string.IsNullOrEmpty(OTP))
-            {
-                yield return new ValidationResult("MasterPasswordHash or OTP must be supplied.");
-            }
-        }
-
-        public bool SuppliedMasterPassword()
-        {
-            return !string.IsNullOrEmpty(MasterPasswordHash);
-        }
-    }
+    public class ApiKeyRequestModel : VerifyPasswordRequestModel { }
 }
